feat: Introduce anime_rpc_bridge developing using rust and native messaging
After a long debate about the development of the rpc bridge, the new decision is now being implemented. This means that the local web server will be completely removed and only the extension can be used to set and control the rich presence. As a background service is still required to communicate with the discord ipc gateway, the aim is to keep this service as minimalistic as possible. To achieve this, the background application is now written in Rust, as this is a language designed for cross-platform compilation, which eliminates dependencies such as pyinstaller or python. Also, communication is now done via the browser's native communication protocol, which is both designed for this case and minimalistic, as it is only reading and writing stdin and stdout.

In this commit, the old Python stuff was completely removed and the basic settings for the native communication part were made.